# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "企業分析エージェント API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/companies": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "List Companies",
        "description": "登録済み企業の一覧を返却する。",
        "operationId": "list_companies_api_companies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "companies"
        ],
        "summary": "Register Company",
        "description": "企業URLを登録し、フルパイプライン（クロール→抽出→分析）を自動開始する。\n\n重複URLの場合は既存レコードを返却する（201ではなく200相当だが冪等性を優先）。",
        "operationId": "register_company_api_companies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterCompanyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/companies/{company_id}": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Get Company",
        "description": "企業詳細を返却する。最新分析結果と直近の実行履歴を含む。",
        "operationId": "get_company_api_companies__company_id__get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/companies/{company_id}/crawl": {
      "post": {
        "tags": [
          "companies"
        ],
        "summary": "Start Crawl",
        "description": "指定企業のクロールジョブ（フルパイプライン）を開始する。",
        "operationId": "start_crawl_api_companies__company_id__crawl_post",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/companies/{company_id}/analysis-runs": {
      "post": {
        "tags": [
          "companies"
        ],
        "summary": "Start Analysis",
        "description": "分析ジョブを開始する。\n\n- force_refresh=true: フルパイプライン（再クロール→抽出→分析）\n- force_refresh=false かつ蓄積データあり: Deep Analysis（テンプレート別再分析）\n- force_refresh=false かつ蓄積データなし: フルパイプライン",
        "operationId": "start_analysis_api_companies__company_id__analysis_runs_post",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartAnalysisRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/companies/{company_id}/analysis-results/latest": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Get Latest Result",
        "description": "指定企業の最新分析結果を返却する。",
        "operationId": "get_latest_result_api_companies__company_id__analysis_results_latest_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/companies/{company_id}/analysis-results": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "List Results",
        "description": "指定企業の分析結果履歴を返却する。",
        "operationId": "list_results_api_companies__company_id__analysis_results_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisHistoryResponse": {
        "properties": {
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/AnalysisResponse"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "results"
        ],
        "title": "AnalysisHistoryResponse"
      },
...
      "AnalysisResponse": {
        "properties": {
          "company_url": {
            "type": "string",
            "title": "Company Url"
          },
          "result_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Id"
          },
          "company_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Id"
          },
          "is_cached": {
            "type": "boolean",
            "title": "Is Cached",
            "default": false
          },
          "analyzed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyzed At"
          },
          "structured": {
            "$ref": "#/components/schemas/StructuredData"
          },
          "summary": {
            "$ref": "#/components/schemas/SummaryData"
          },
          "scores": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScoreData"
              },
              {
                "type": "null"
              }
            ]
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceInfo"
            },
            "type": "array",
            "title": "Sources"
          },
          "raw_sources": {
            "items": {
              "$ref": "#/components/schemas/RawSource"
            },
            "type": "array",
            "title": "Raw Sources"
          },
          "markdown_page": {
            "type": "string",
            "title": "Markdown Page",
            "default": ""
          },
          "diff_report": {
            "type": "string",
            "title": "Diff Report",
            "default": ""
          },
          "template": {
            "type": "string",
            "title": "Template",
            "default": "general"
          }
        },
        "type": "object",
        "required": [
          "company_url",
          "structured",
          "summary",
          "sources"
        ],
        "title": "AnalysisResponse"
      },
      "CompanyDetailResponse": {
        "properties": {
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id"
          },
          "primary_url": {
            "type": "string",
            "title": "Primary Url"
          },
          "normalized_url": {
            "type": "string",
            "title": "Normalized Url"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "last_page_crawl_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Page Crawl At"
          },
          "last_analyzed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Analyzed At"
          },
          "analysis_count": {
            "type": "integer",
            "title": "Analysis Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "latest_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "recent_runs": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array",
            "title": "Recent Runs"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "primary_url",
          "normalized_url",
          "status",
          "analysis_count",
          "created_at"
        ],
        "title": "CompanyDetailResponse"
      },
      "CompanyListResponse": {
        "properties": {
          "companies": {
            "items": {
              "$ref": "#/components/schemas/CompanyResponse"
            },
            "type": "array",
            "title": "Companies"
          }
        },
        "type": "object",
        "required": [
          "companies"
        ],
        "title": "CompanyListResponse"
      },
...
      "CompanyResponse": {
        "properties": {
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id"
          },
          "primary_url": {
            "type": "string",
            "title": "Primary Url"
          },
          "normalized_url": {
            "type": "string",
            "title": "Normalized Url"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "last_page_crawl_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Page Crawl At"
          },
          "last_analyzed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Analyzed At"
          },
          "analysis_count": {
            "type": "integer",
            "title": "Analysis Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "company_id",
          "primary_url",
          "normalized_url",
          "status",
          "analysis_count",
          "created_at"
        ],
        "title": "CompanyResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RegisterCompanyRequest": {
        "properties": {
          "url": {
            "type": "string",
            "minLength": 1,
            "pattern": "^https?://.*",
            "title": "Url",
            "description": "登録する企業のURL（例: https://www.toyota.co.jp/）"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "RegisterCompanyRequest"
      },
...
      "RunStatusResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "ジョブ状態: pending, running, completed, failed"
          },
          "run_type": {
            "type": "string",
            "title": "Run Type"
          },
          "template": {
            "type": "string",
            "title": "Template"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "result_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Id"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "company_id",
          "status",
          "run_type",
          "template"
        ],
        "title": "RunStatusResponse"
      },
      "RunSummary": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "run_type": {
            "type": "string",
            "title": "Run Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "template": {
            "type": "string",
            "title": "Template"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "result_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Id"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "run_type",
          "status",
          "template",
          "started_at",
          "completed_at",
          "duration_ms",
          "error_message",
          "result_id"
        ],
        "title": "RunSummary"
      },
      "ScoreData": {
        "properties": {
          "financial_health": {
            "$ref": "#/components/schemas/ScoreItem"
          },
          "growth_potential": {
            "$ref": "#/components/schemas/ScoreItem"
          },
          "competitive_edge": {
            "$ref": "#/components/schemas/ScoreItem"
          },
          "risk_level": {
            "$ref": "#/components/schemas/ScoreItem"
          },
          "info_transparency": {
            "$ref": "#/components/schemas/ScoreItem"
          }
        },
        "type": "object",
        "title": "ScoreData"
      },
...
      "StartAnalysisRequest": {
        "properties": {
          "template": {
            "type": "string",
            "title": "Template",
            "description": "分析テンプレート",
            "default": "general"
          },
          "force_refresh": {
            "type": "boolean",
            "title": "Force Refresh",
            "description": "キャッシュを無視して再クロール・再分析する",
            "default": false
          }
        },
        "type": "object",
        "title": "StartAnalysisRequest"
      },
      "StructuredData": {
        "properties": {
          "company_profile": {
            "$ref": "#/components/schemas/CompanyProfile"
          },
          "business_domains": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Business Domains"
          },
          "products": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Products"
          },
          "financials": {
            "$ref": "#/components/schemas/Financials"
          },
          "news": {
            "items": {
              "$ref": "#/components/schemas/NewsItem"
            },
            "type": "array",
            "title": "News"
          },
          "risks": {
            "items": {
              "$ref": "#/components/schemas/RiskItem"
            },
            "type": "array",
            "title": "Risks"
          }
        },
        "type": "object",
        "title": "StructuredData"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}